requirements.txt
anyio==4.1.0
apturl==0.5.2
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.11.0
blinker==1.4
Brlapi==0.7.0
build==1.0.3
CacheControl==0.13.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
cleo==2.1.0
click==8.1.7
colorama==0.4.3
command-not-found==0.3
crashtest==0.4.1
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.7
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.7
django-cors-headers==4.3.1
django-crontab==0.7.1
django-filter==23.4
djangorestframework==3.14.0
dulwich==0.21.6
entrypoints==0.3
exceptiongroup==1.2.0
fastjsonschema==2.19.0
filelock==3.13.1
h11==0.14.0
httpcore==1.0.2
httplib2==0.14.0
httpx==0.25.2
idna==2.8
importlib-metadata==6.9.0
importlib-resources==6.1.1
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Markdown==3.5.1
more-itertools==10.1.0
msgpack==1.0.7
mypy-extensions==1.0.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==23.2
pathspec==0.11.2
pbr==6.0.0
pexpect==4.9.0
Pillow==7.0.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
protobuf==3.6.1
psycopg==3.1.13
psycopg2==2.9.9
ptyprocess==0.7.0
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyproject_hooks==1.0.0
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
rapidfuzz==3.5.2
reportlab==3.5.34
requests==2.31.0
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
SecretStorage==3.3.3
shellingham==1.5.4
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.3.0
sqlparse==0.4.4
ssh-import-id==5.10
systemd-python==234
testresources==2.0.1
tomli==2.0.1
tomlkit==0.12.3
trove-classifiers==2023.11.29
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.25.0
wadllib==1.3.3
xkit==0.0.0
zipp==3.17.0
